requirements.txt
fastapi==0.100.0
hypercorn==0.14.4
openai==0.28.1
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
langchain==0.0.305
requests==2.31.0
bs4==0.0.1
